Restructure middlewares into folder
